Move OGI fix on FieldObjectEntities to always-on patches
